wip: set up fetch to real API
